requirements.txt
keyring==9.3.1
pycrypto==2.6.1
tabulate==0.7.7
